// src/app/models/invoice/invoice-criteria-view-model.ts
/**
 * Fee
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */


/**
 * Invoice Criteria ViewModel
 */
export interface InvoiceCriteriaViewModel { 
    /**
     * FeeTermId
     */
    feeTermId?: number;
    /**
     * BoardId
     */
    boardId?: number;
    /**
     * ClassId
     */
    classId?: number;
    /**
     * ClassId
     */
    sectionId?: number;
}